Viewer UI for firmware update
