+++ common/ux-window.cpp
@@ -33,6 +33,7 @@ namespace rs2
         config_file::instance().set_default(configurations::viewer::log_to_console, true);
         config_file::instance().set_default(configurations::viewer::log_to_file, false);
         config_file::instance().set_default(configurations::viewer::log_severity, 2);
+        config_file::instance().set_default(configurations::viewer::metric_system, true);
 
         config_file::instance().set_default(configurations::record::compression_mode, 2); // Let the device decide
         config_file::instance().set_default(configurations::record::default_path, get_folder_path(special_folder::user_documents));
@@ -363,6 +364,86 @@ namespace rs2
         _on_load_message.push_back(msg);
     }
 
+    void ux_window::imgui_config_pop()
+    {
+        ImGui::PopFont();
+        ImGui::End();
+
+        ImGui::PopStyleColor(3);
+        ImGui::PopStyleVar(2);
+        end_frame();
+
+        glPopMatrix();
+    }
+
+    void ux_window::imgui_config_push()
+    {
+        glPushMatrix();
+        glViewport(0, 0, _fb_width, _fb_height);
+        glClearColor(0.036f, 0.044f, 0.051f, 1.f);
+        glClear(GL_COLOR_BUFFER_BIT);
+
+        glLoadIdentity();
+        glOrtho(0, _width, _height, 0, -1, +1);
+
+        // Fade-in the logo
+        auto opacity = smoothstep(float(_splash_timer.elapsed_ms()), 100.f, 2500.f);
+        auto ox = 0.7f - smoothstep(float(_splash_timer.elapsed_ms()), 200.f, 1900.f) * 0.4f;
+        auto oy = 0.5f;
+        auto power = std::sin(smoothstep(float(_splash_timer.elapsed_ms()), 150.f, 2200.f) * 3.14f) * 0.96f;
+
+        if (_use_glsl_render)
+        {
+            auto shader = ((splash_screen_shader*)&_2d_vis->get_shader());
+            shader->begin();
+            shader->set_power(power);
+            shader->set_ray_center(float2{ ox, oy });
+            shader->end();
+            _2d_vis->draw_texture(_splash_tex.get_gl_handle(), opacity);
+        }
+        else
+        {
+            _splash_tex.show({ 0.f,0.f,float(_width),float(_height) }, opacity);
+        }
+
+        std::string hourglass = u8"\uf250";
+        static periodic_timer every_200ms(std::chrono::milliseconds(200));
+        bool do_200ms = every_200ms;
+        if (_query_devices && do_200ms)
+        {
+            _missing_device = rs2::context().query_devices(RS2_PRODUCT_LINE_ANY).size() == 0;
+            _hourglass_index = (_hourglass_index + 1) % 5;
+
+            if (!_missing_device)
+            {
+                _dev_stat_message = u8"\uf287 RealSense device detected.";
+                _query_devices = false;
+            }
+        }
+
+        hourglass[2] += _hourglass_index;
+
+        auto flags = ImGuiWindowFlags_NoResize |
+            ImGuiWindowFlags_NoMove |
+            ImGuiWindowFlags_NoCollapse |
+            ImGuiWindowFlags_NoTitleBar;
+
+        auto text_color = light_grey;
+        text_color.w = opacity;
+        ImGui::PushStyleColor(ImGuiCol_Text, text_color);
+        ImGui::PushStyleColor(ImGuiCol_TextSelectedBg, white);
+        ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, { 5, 5 });
+        ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding, 1);
+        ImGui::PushStyleColor(ImGuiCol_WindowBg, transparent);
+        ImGui::SetNextWindowPos({ (float)_width / 2 - 150, (float)_height / 2 + 70 });
+
+        ImGui::SetNextWindowSize({ (float)_width, (float)_height });
+        ImGui::Begin("Splash Screen Banner", nullptr, flags);
+        ImGui::PushFont(_font_18);
+
+        ImGui::Text("%s   Loading %s...", hourglass.c_str(), _title_str.c_str());
+    }
+
     // Check that the graphic subsystem is valid and start a new frame
     ux_window::operator bool()
     {
@@ -392,9 +473,11 @@ namespace rs2
 
         auto res = !glfwWindowShouldClose(_win);
 
+
         if (_first_frame)
         {
             assert(!_first_load.joinable()); // You must call to reset() before initiate new thread
+
 
             _first_load = std::thread([&]() {
                 while (_keep_alive && !_app_ready)
@@ -424,70 +507,8 @@ namespace rs2
                 _is_ui_aligned = is_gui_aligned(_win);
                 _first_frame = false;
             }
-            glPushMatrix();
-            glViewport(0, 0, _fb_width, _fb_height);
-            glClearColor(0.036f, 0.044f, 0.051f, 1.f);
-            glClear(GL_COLOR_BUFFER_BIT);
 
-            glLoadIdentity();
-            glOrtho(0, _width, _height, 0, -1, +1);
-
-            // Fade-in the logo
-            auto opacity = smoothstep(float(_splash_timer.elapsed_ms()), 100.f, 2500.f);
-            auto ox = 0.7f - smoothstep(float(_splash_timer.elapsed_ms()), 200.f, 1900.f) * 0.4f;
-            auto oy = 0.5f;
-            auto power = std::sin(smoothstep(float(_splash_timer.elapsed_ms()), 150.f, 2200.f) * 3.14f) * 0.96f;
-
-            if (_use_glsl_render)
-            {
-                auto shader = ((splash_screen_shader*)&_2d_vis->get_shader());
-                shader->begin();
-                shader->set_power(power);
-                shader->set_ray_center(float2{ox, oy});
-                shader->end();
-                _2d_vis->draw_texture(_splash_tex.get_gl_handle(), opacity);
-            }
-            else
-            {
-                _splash_tex.show({ 0.f,0.f,float(_width),float(_height) }, opacity);
-            }
-
-            std::string hourglass = u8"\uf250";
-            static periodic_timer every_200ms(std::chrono::milliseconds(200));
-            bool do_200ms = every_200ms;
-            if (_query_devices && do_200ms)
-            {
-                _missing_device = rs2::context().query_devices(RS2_PRODUCT_LINE_ANY).size() == 0;
-                _hourglass_index = (_hourglass_index + 1) % 5;
-
-                if (!_missing_device)
-                {
-                    _dev_stat_message = u8"\uf287 RealSense device detected.";
-                    _query_devices = false;
-                }
-            }
-
-            hourglass[2] += _hourglass_index;
-
-            auto flags = ImGuiWindowFlags_NoResize |
-                ImGuiWindowFlags_NoMove |
-                ImGuiWindowFlags_NoCollapse |
-                ImGuiWindowFlags_NoTitleBar;
-
-            auto text_color = light_grey;
-            text_color.w = opacity;
-            ImGui::PushStyleColor(ImGuiCol_Text, text_color);
-            ImGui::PushStyleColor(ImGuiCol_TextSelectedBg, white);
-            ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, { 5, 5 });
-            ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding, 1);
-            ImGui::PushStyleColor(ImGuiCol_WindowBg, transparent);
-            ImGui::SetNextWindowPos({ (float)_width / 2 - 150, (float)_height / 2 + 70 });
-            
-            ImGui::SetNextWindowSize({ (float)_width, (float)_height });
-            ImGui::Begin("Splash Screen Banner", nullptr, flags);
-            ImGui::PushFont(_font_18);
-
-            ImGui::Text("%s   Loading %s...", hourglass.c_str(), _title_str.c_str());
+            imgui_config_push();
 
             {
                 std::lock_guard<std::mutex> lock(_on_load_message_mtx);
@@ -509,15 +530,8 @@ namespace rs2
                 }
             }
 
-            ImGui::PopFont();
-            ImGui::End();
-            
-            ImGui::PopStyleColor(3);
-            ImGui::PopStyleVar(2);
+            imgui_config_pop();
 
-            end_frame();
-
-            glPopMatrix();
 
             // Yield the CPU
             std::this_thread::sleep_for(std::chrono::milliseconds(10));
