+++ tools/realsense-viewer/CMakeLists.txt
@@ -8,92 +8,92 @@ project(RealsenseToolsRealSenseViewer)
 set(THREADS_PREFER_PTHREAD_FLAG ON)
 find_package(Threads REQUIRED)
 
+include_directories(${CMAKE_BINARY_DIR})
+
+
+include(../../common/CMakeLists.txt)
+
 if(BUILD_GRAPHICAL_EXAMPLES)
+set(RS_VIEWER_CPP
+        ${COMMON_SRC}
+        realsense-viewer.cpp
+        ../../third-party/imgui/imgui.cpp
+        ../../third-party/imgui/imgui_draw.cpp
+        ../../third-party/imgui/imgui_impl_glfw.cpp
+        ../../third-party/imgui/imgui-fonts-karla.hpp
+        ../../third-party/imgui/imgui-fonts-fontawesome.hpp
+        ../../common/realsense-ui-advanced-mode.h
+        ../../common/rendering.h
+        ../../common/model-views.h
+        ../../common/model-views.cpp
+        ../../common/ux-window.h
+        ../../common/ux-window.cpp
+        ../../common/ux-alignment.cpp
+        ../../common/ux-alignment.h
+        ../../third-party/glad/glad.c
+        ../../third-party/tinyfiledialogs/tinyfiledialogs.c
+        ../../third-party/tinyfiledialogs/tinyfiledialogs.h
+        ../../common/opengl3.cpp
+        ../../common/opengl3.h
+        ../../common/rs-config.h
+        ../../common/rs-config.cpp
+        ../../common/os.h
+        ../../common/os.cpp
+)
 
-if(APPLE)
-    add_definitions(-DNOC_FILE_DIALOG_OSX)
-elseif(UNIX)
-    find_package(PkgConfig REQUIRED)
-    pkg_check_modules(GTK3 REQUIRED gtk+-3.0)
+# config-ui
+if(WIN32)
+    add_executable(realsense-viewer WIN32
+            ${RS_VIEWER_CPP} ${CMAKE_CURRENT_SOURCE_DIR}/res/resource.h
+            ${CMAKE_CURRENT_SOURCE_DIR}/res/realsense-viewer.rc
+            ../../common/windows-app-bootstrap.cpp)
 
-    include_directories(${GTK3_INCLUDE_DIRS})
-    link_directories(${GTK3_LIBRARY_DIRS})
-    add_definitions(${GTK3_CFLAGS_OTHER})
-
-    add_definitions(-DNOC_FILE_DIALOG_GTK)
-elseif(WIN32)
-    add_definitions(-DNOC_FILE_DIALOG_WIN32)
-endif()
-
-    set(RS_VIEWER_CPP
-            realsense-viewer.cpp
-            ../../third-party/imgui/imgui.cpp
-            ../../third-party/imgui/imgui_draw.cpp
-            ../../third-party/imgui/imgui_impl_glfw.cpp
-            ../../third-party/imgui/imgui-fonts-karla.hpp
-            ../../third-party/imgui/imgui-fonts-fontawesome.hpp
-            ../../common/rendering.h
-            ../../common/realsense-ui-advanced-mode.h
-            ../../common/model-views.h
-            ../../common/model-views.cpp
-            ../../common/ux-window.h
-            ../../common/ux-window.cpp
-            ../../common/ux-alignment.cpp
-            ../../common/ux-alignment.h
-            ../../third-party/glad/glad.c
-            ../../common/opengl3.cpp
-            ../../common/opengl3.h
-            ../../common/rs-config.h
-            ../../common/rs-config.cpp
-            ../../common/os.h
-            ../../common/os.cpp
+    source_group("3rd Party" FILES
+            ../../third-party/tinyfiledialogs/tinyfiledialogs.c
+            ../../third-party/tinyfiledialogs/tinyfiledialogs.h
+        ../../third-party/imgui/imgui.cpp
+        ../../third-party/imgui/imgui_draw.cpp
+        ../../third-party/imgui/imgui_impl_glfw.cpp
+        ../../third-party/imgui/imgui-fonts-karla.hpp
+        ../../third-party/imgui/imgui-fonts-fontawesome.hpp
         )
 
-    # config-ui
-    if(WIN32)
-        add_executable(realsense-viewer WIN32
-                ${RS_VIEWER_CPP} ${CMAKE_CURRENT_SOURCE_DIR}/res/resource.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/res/realsense-viewer.rc
-                ../../common/windows-app-bootstrap.cpp)
+    source_group("Resources" FILES
+            ${CMAKE_CURRENT_SOURCE_DIR}/res/resource.h
+            ${CMAKE_CURRENT_SOURCE_DIR}/res/realsense-viewer.rc)
 
-        source_group("3rd Party" FILES
-            ../../third-party/imgui/imgui.cpp
-            ../../third-party/imgui/imgui_draw.cpp
-            ../../third-party/imgui/imgui_impl_glfw.cpp
-            ../../third-party/imgui/imgui-fonts-karla.hpp
-            ../../third-party/imgui/imgui-fonts-fontawesome.hpp
-            )
+    include_directories(realsense-viewer ../../third-party/imgui 
+    ../../third-party/glad ../../common
+    ../../third-party ${CMAKE_CURRENT_SOURCE_DIR}/res/ ../../third-party/tinyfiledialogs)
+else()
+    add_executable(realsense-viewer ${RS_VIEWER_CPP})
+    include_directories(realsense-viewer ../../third-party/imgui 
+    ../../third-party/glad ../../common ../../third-party ../../third-party/tinyfiledialogs)
+endif()
 
-        source_group("Resources" FILES
-                ${CMAKE_CURRENT_SOURCE_DIR}/res/resource.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/res/realsense-viewer.rc)
+set_property(TARGET realsense-viewer PROPERTY CXX_STANDARD 11)
 
-        include_directories(realsense-viewer ../../third-party/imgui 
-        ../../third-party/glad ../../common
-        ../../third-party ${CMAKE_CURRENT_SOURCE_DIR}/res/)
-    else()
-        add_executable(realsense-viewer ${RS_VIEWER_CPP})
-        include_directories(realsense-viewer ../../third-party/imgui 
-        ../../third-party/glad ../../common ../../third-party)
-    endif()
-    set_property(TARGET realsense-viewer PROPERTY CXX_STANDARD 11)
+target_link_libraries(realsense-viewer ${DEPENDENCIES}
+                                        ${GTK3_LIBRARIES}
+                                        Threads::Threads
+                                        realsense2-gl
+                                        )
 
-    target_link_libraries(realsense-viewer ${DEPENDENCIES}
-                                           ${GTK3_LIBRARIES}
-                                           Threads::Threads
-                                           realsense2-gl
-                                           )
+if (IMPORT_UVC_FW)
+    add_definitions(-DINTERNAL_FW)
+    target_link_libraries(realsense-viewer uvc_fw)
+endif()
 
-    set_target_properties (realsense-viewer PROPERTIES
-        FOLDER Tools
-    )
+set_target_properties (realsense-viewer PROPERTIES
+    FOLDER Tools
+)
 
-    install(
-        TARGETS
+install(
+    TARGETS
 
-        realsense-viewer
+    realsense-viewer
 
-        RUNTIME DESTINATION
-        ${CMAKE_INSTALL_BINDIR}
-    )
+    RUNTIME DESTINATION
+    ${CMAKE_INSTALL_BINDIR}
+)
 endif()
