+++ tools/realsense-viewer/realsense-viewer.cpp
@@ -2,9 +2,10 @@
 // Copyright(c) 2017 Intel Corporation. All Rights Reserved.
 
 #include <librealsense2/rs.hpp>
-#include "model-views.h"
+#include "viewer.h"
 #include "os.h"
 #include "ux-window.h"
+#include "fw-update-helper.h"
 
 #include <cstdarg>
 #include <thread>
@@ -19,13 +20,18 @@
 
 #include <imgui_internal.h>
 
-// We use NOC file helper function for cross-platform file dialogs
-#include <noc_file_dialog.h>
+#ifdef INTERNAL_FW
+#include "common/fw/D4XX_FW_Image.h"
+#include "common/fw/SR3XX_FW_Image.h"
+#else
+#define FW_D4XX_FW_IMAGE_VERSION ""
+#define FW_SR3XX_FW_IMAGE_VERSION ""
+#endif // INTERNAL_FW
 
 using namespace rs2;
 using namespace rs400;
 
-void add_playback_device(context& ctx, std::shared_ptr<std::vector<device_model>> device_models, 
+void add_playback_device(context& ctx, device_models_list& device_models, 
     std::string& error_message, viewer_model& viewer_model, const std::string& file)
 {
     bool was_loaded = false;
@@ -34,30 +40,30 @@ void add_playback_device(context& ctx, std::shared_ptr<std::vector<device_model>
     {
         auto dev = ctx.load_device(file);
         was_loaded = true;
-        device_models->emplace_back(dev, error_message, viewer_model); //Will cause the new device to appear in the left panel
+        device_models.emplace_back(new device_model(dev, error_message, viewer_model)); //Will cause the new device to appear in the left panel
         if (auto p = dev.as<playback>())
         {
             auto filename = p.file_name();
-            p.set_status_changed_callback([&viewer_model, device_models, filename](rs2_playback_status status)
+            p.set_status_changed_callback([&viewer_model, &device_models, filename](rs2_playback_status status)
             {
                 if (status == RS2_PLAYBACK_STATUS_STOPPED)
                 {
-                    auto it = std::find_if(device_models->begin(), device_models->end(),
-                        [&](const device_model& dm) {
-                        if (auto p = dm.dev.as<playback>())
+                    auto it = std::find_if(device_models.begin(), device_models.end(),
+                        [&](const std::unique_ptr<device_model>& dm) {
+                        if (auto p = dm->dev.as<playback>())
                             return p.file_name() == filename;
                         return false;
                     });
-                    if (it != device_models->end())
+                    if (it != device_models.end())
                     {
-                        auto subs = it->subdevices;
-                        if (it->_playback_repeat)
+                        auto subs = (*it)->subdevices;
+                        if ((*it)->_playback_repeat)
                         {
                             //Calling from different since playback callback is from reading thread
                             std::thread{ [subs, &viewer_model, it]()
                             {
-                                if(!it->dev_syncer)
-                                    it->dev_syncer = viewer_model.syncer->create_syncer();
+                                if(!(*it)->dev_syncer)
+                                    (*it)->dev_syncer = viewer_model.syncer->create_syncer();
 
                                 for (auto&& sub : subs)
                                 {
@@ -65,7 +71,7 @@ void add_playback_device(context& ctx, std::shared_ptr<std::vector<device_model>
                                     {
                                         auto profiles = sub->get_selected_profiles();
 
-                                        sub->play(profiles, viewer_model, it->dev_syncer);
+                                        sub->play(profiles, viewer_model, (*it)->dev_syncer);
                                     }
                                 }
                             } }.detach();
@@ -105,121 +111,138 @@ void add_playback_device(context& ctx, std::shared_ptr<std::vector<device_model>
 // This function is called every frame
 // If between the frames there was an asyncronous connect/disconnect event
 // the function will pick up on this and add the device to the viewer
-void refresh_devices(std::mutex& m,
+bool refresh_devices(std::mutex& m,
     context& ctx,
     device_changes& devices_connection_changes,
     std::vector<device>& current_connected_devices,
     std::vector<std::pair<std::string, std::string>>& device_names,
-    std::shared_ptr<std::vector<device_model>> device_models,
+    device_models_list& device_models,
     viewer_model& viewer_model,
     std::string& error_message)
 {
     event_information info({}, {});
-    if (devices_connection_changes.try_get_next_changes(info))
+    if (!devices_connection_changes.try_get_next_changes(info))
+        return false;
+    try
     {
-        try
+        auto prev_size = current_connected_devices.size();
+
+        //Remove disconnected
+        auto dev_itr = begin(current_connected_devices);
+        while (dev_itr != end(current_connected_devices))
         {
-            auto prev_size = current_connected_devices.size();
-
-            //Remove disconnected
-            auto dev_itr = begin(current_connected_devices);
-            while (dev_itr != end(current_connected_devices))
+            auto dev = *dev_itr;
+            if (info.was_removed(dev))
             {
-                auto dev = *dev_itr;
-                if (info.was_removed(dev))
+                //Notify change
+                viewer_model.not_model.add_notification({ get_device_name(dev).first + " Disconnected\n",
+                    RS2_LOG_SEVERITY_INFO, RS2_NOTIFICATION_CATEGORY_UNKNOWN_ERROR });
+
+                //Remove from devices
+                auto dev_model_itr = std::find_if(begin(device_models), end(device_models),
+                    [&](const std::unique_ptr<device_model>& other) { return get_device_name(other->dev) == get_device_name(dev); });
+
+                if (dev_model_itr != end(device_models))
                 {
-                    //Notify change
-                    viewer_model.not_model.add_notification({ get_device_name(dev).first + " Disconnected\n",
-                        0, RS2_LOG_SEVERITY_INFO, RS2_NOTIFICATION_CATEGORY_UNKNOWN_ERROR });
+                    for (auto&& s : (*dev_model_itr)->subdevices)
+                        s->streaming = false;
 
-                    //Remove from devices
-                    auto dev_model_itr = std::find_if(begin(*device_models), end(*device_models),
-                        [&](const device_model& other) { return get_device_name(other.dev) == get_device_name(dev); });
+                    dev_model_itr->reset();
+                    device_models.erase(dev_model_itr);
 
-                    if (dev_model_itr != end(*device_models))
+                    if (device_models.size() == 0)
                     {
-                        for (auto&& s : dev_model_itr->subdevices)
-                            s->streaming = false;
+                        viewer_model.ppf.depth_stream_active = false;
 
-                        dev_model_itr->reset();
-                        device_models->erase(dev_model_itr);
+                        // Stopping post processing filter rendering thread in case of disconnection
+                        viewer_model.ppf.stop();
+                    }
+                }
+                auto dev_name_itr = std::find(begin(device_names), end(device_names), get_device_name(dev));
+                if (dev_name_itr != end(device_names))
+                    device_names.erase(dev_name_itr);
 
-                        if(device_models->size() == 0)
+                dev_itr = current_connected_devices.erase(dev_itr);
+                continue;
+            }
+            ++dev_itr;
+        }
+
+        //Add connected
+        static bool initial_refresh = true;
+        for (auto dev : info.get_new_devices())
+        {
+            auto dev_descriptor = get_device_name(dev);
+            device_names.push_back(dev_descriptor);
+
+            bool added = false;
+            if (device_models.size() == 0 &&
+                dev.supports(RS2_CAMERA_INFO_NAME) && std::string(dev.get_info(RS2_CAMERA_INFO_NAME)) != "Platform Camera")
+            {
+                device_models.emplace_back(new device_model(dev, error_message, viewer_model));
+                viewer_model.not_model.add_log(to_string() << (*device_models.rbegin())->dev.get_info(RS2_CAMERA_INFO_NAME) << " was selected as a default device");
+                added = true;
+            }
+
+            if (!initial_refresh)
+            {
+                if (added || dev.is<playback>())
+                viewer_model.not_model.add_notification({ dev_descriptor.first + " Connected\n",
+                        RS2_LOG_SEVERITY_INFO, RS2_NOTIFICATION_CATEGORY_UNKNOWN_ERROR });
+                else
+                    viewer_model.not_model.add_notification({ dev_descriptor.first + " Connected\n",
+                        RS2_LOG_SEVERITY_INFO, RS2_NOTIFICATION_CATEGORY_UNKNOWN_ERROR }, 
+                        [&device_models, &viewer_model, &error_message, dev]{
+                            auto device = dev;
+                            device_models.emplace_back(
+                                new device_model(device, error_message, viewer_model));
+                        });
+            }
+
+            current_connected_devices.push_back(dev);
+            for (auto&& s : dev.query_sensors())
+            {
+                s.set_notifications_callback([&, dev_descriptor](const notification& n)
+                {
+                    if (n.get_category() == RS2_NOTIFICATION_CATEGORY_HARDWARE_EVENT)
+                    {
+                        auto data = n.get_serialized_data();
+                        if (!data.empty())
                         {
-                            viewer_model.ppf.depth_stream_active = false;
+                            auto dev_model_itr = std::find_if(begin(device_models), end(device_models),
+                                [&](const std::unique_ptr<device_model>& other) 
+                                { return get_device_name(other->dev) == dev_descriptor; });
 
-                            // Stopping post processing filter rendering thread in case of disconnection
-                            viewer_model.ppf.stop();
+                            if (dev_model_itr == end(device_models))
+                                return;
+
+                            (*dev_model_itr)->handle_hardware_events(data);
                         }
                     }
-                    auto dev_name_itr = std::find(begin(device_names), end(device_names), get_device_name(dev));
-                    if (dev_name_itr != end(device_names))
-                        device_names.erase(dev_name_itr);
-
-                    dev_itr = current_connected_devices.erase(dev_itr);
-                    continue;
-                }
-                ++dev_itr;
+                    viewer_model.not_model.add_notification({ n.get_description(), n.get_severity(), n.get_category() });
+                });
             }
 
-            //Add connected
-            static bool initial_refresh = true;
-            for (auto dev : info.get_new_devices())
-            {
-                auto dev_descriptor = get_device_name(dev);
-                device_names.push_back(dev_descriptor);
-                if (!initial_refresh)
-                    viewer_model.not_model.add_notification({ dev_descriptor.first + " Connected\n",
-                        0, RS2_LOG_SEVERITY_INFO, RS2_NOTIFICATION_CATEGORY_UNKNOWN_ERROR });
-
-                current_connected_devices.push_back(dev);
-                for (auto&& s : dev.query_sensors())
-                {
-                    s.set_notifications_callback([&, dev_descriptor](const notification& n)
-                    {
-                        if (n.get_category() == RS2_NOTIFICATION_CATEGORY_HARDWARE_EVENT)
-                        {
-                            auto data = n.get_serialized_data();
-                            if (!data.empty())
-                            {
-                                auto dev_model_itr = std::find_if(begin(*device_models), end(*device_models),
-                                    [&](const device_model& other) { return get_device_name(other.dev) == dev_descriptor; });
-
-                                if (dev_model_itr == end(*device_models))
-                                    return;
-
-                                dev_model_itr->handle_hardware_events(data);
-                            }
-                        }
-                        viewer_model.not_model.add_notification({ n.get_description(), n.get_timestamp(), n.get_severity(), n.get_category() });
-                    });
-                }
-
-                if (device_models->size() == 0 &&
-                    dev.supports(RS2_CAMERA_INFO_NAME) && std::string(dev.get_info(RS2_CAMERA_INFO_NAME)) != "Platform Camera")
-                {
-                    device_models->emplace_back(dev, error_message, viewer_model);
-                    viewer_model.not_model.add_log(to_string() << device_models->rbegin()->dev.get_info(RS2_CAMERA_INFO_NAME) << " was selected as a default device");
-                }
-            }
-            initial_refresh = false;
+            
         }
-        catch (const error& e)
-        {
-            error_message = error_to_string(e);
-        }
-        catch (const std::exception& e)
-        {
-            error_message = e.what();
-        }
-        catch (...)
-        {
-            error_message = "Unknown error";
-        }
+        initial_refresh = false;
     }
+    catch (const error& e)
+    {
+        error_message = error_to_string(e);
+    }
+    catch (const std::exception& e)
+    {
+        error_message = e.what();
+    }
+    catch (...)
+    {
+        error_message = "Unknown error";
+    }
+    return true;
 }
 
-int main(int argv, const char** argc) try
+int main(int argc, const char** argv) try
 {
     rs2::log_to_console(RS2_LOG_SEVERITY_WARN);
 
@@ -233,10 +256,11 @@ int main(int argv, const char** argc) try
     std::string error_message{ "" };
     std::string label{ "" };
 
-    std::shared_ptr<std::vector<device_model>> device_models = std::make_shared<std::vector<device_model>>();
+    std::shared_ptr<device_models_list> device_models = std::make_shared<device_models_list>();
     device_model* device_to_remove = nullptr;
 
     viewer_model viewer_model;
+    viewer_model.ctx = ctx;
 
     std::vector<device> connected_devs;
     std::mutex m;
@@ -244,24 +268,24 @@ int main(int argv, const char** argc) try
     window.on_file_drop = [&](std::string filename)
     {
         std::string error_message{};
-        add_playback_device(ctx, device_models, error_message, viewer_model, filename);
+        add_playback_device(ctx, *device_models, error_message, viewer_model, filename);
         if (!error_message.empty())
         {
             viewer_model.not_model.add_notification({ error_message,
-                0, RS2_LOG_SEVERITY_ERROR, RS2_NOTIFICATION_CATEGORY_UNKNOWN_ERROR });
+                RS2_LOG_SEVERITY_ERROR, RS2_NOTIFICATION_CATEGORY_UNKNOWN_ERROR });
         }
     };
 
-    for (int i = 1; i < argv; i++)
+    for (int i = 1; i < argc; i++)
     {
         try
         {
-            const char* arg = argc[i];
+            const char* arg = argv[i];
             std::ifstream file(arg);
             if (!file.good())
                 continue;
 
-            add_playback_device(ctx, device_models, error_message, viewer_model, arg);
+            add_playback_device(ctx, *device_models, error_message, viewer_model, arg);
         }
         catch (const rs2::error& e)
         {
@@ -275,20 +299,17 @@ int main(int argv, const char** argc) try
 
     window.on_load = [&]()
     {
-        refresh_devices(m, ctx, devices_connection_changes, connected_devs, 
-            device_names, device_models, viewer_model, error_message);
+        refresh_devices(m, ctx, devices_connection_changes, connected_devs,
+            device_names, *device_models, viewer_model, error_message);
         return true;
     };
+
 
     // Closing the window
     while (window)
     {
-        if (!window.is_ui_aligned())
-        {
-            viewer_model.popup_if_ui_not_aligned(window.get_font());
-        }
-        refresh_devices(m, ctx, devices_connection_changes, connected_devs, 
-            device_names, device_models, viewer_model, error_message);
+        auto device_changed = refresh_devices(m, ctx, devices_connection_changes, connected_devs, 
+            device_names, *device_models, viewer_model, error_message);
 
         auto output_height = viewer_model.get_output_height();
 
@@ -320,12 +341,13 @@ int main(int argv, const char** argc) try
             ImGui::OpenPopup("select");
 
         auto new_devices_count = device_names.size() + 1;
+
         for (auto&& dev_model : *device_models)
         {
             auto connected_devs_itr = std::find_if(begin(connected_devs), end(connected_devs),
-                [&](const device& d) { return get_device_name(d) == get_device_name(dev_model.dev); });
+                [&](const device& d) { return get_device_name(d) == get_device_name(dev_model->dev); });
 
-            if (connected_devs_itr != end(connected_devs) || dev_model.dev.is<playback>())
+            if (connected_devs_itr != end(connected_devs) || dev_model->dev.is<playback>())
                 new_devices_count--;
         }
 
@@ -340,7 +362,7 @@ int main(int argv, const char** argc) try
             {
                 bool skip = false;
                 for (auto&& dev_model : *device_models)
-                    if (get_device_name(dev_model.dev) == device_names[i]) skip = true;
+                    if (get_device_name(dev_model->dev) == device_names[i]) skip = true;
                 if (skip) continue;
 
                 if (ImGui::Selectable(device_names[i].first.c_str(), false, ImGuiSelectableFlags_SpanAllColumns)/* || switch_to_newly_loaded_device*/)
@@ -348,7 +370,7 @@ int main(int argv, const char** argc) try
                     try
                     {
                         auto dev = connected_devs[i];
-                        device_models->emplace_back(dev, error_message, viewer_model);
+                        device_models->emplace_back(new device_model(dev, error_message, viewer_model));
                     }
                     catch (const error& e)
                     {
@@ -383,7 +405,7 @@ int main(int argv, const char** argc) try
             {
                 if (auto ret = file_dialog_open(open_file, "ROS-bag\0*.bag\0", NULL, NULL))
                 {
-                    add_playback_device(ctx, device_models, error_message, viewer_model, ret);
+                    add_playback_device(ctx, *device_models, error_message, viewer_model, ret);
                 }
             }
             ImGui::NextColumn();
@@ -428,32 +450,20 @@ int main(int argv, const char** argc) try
 
             for (auto&& dev_model : *device_models)
             {
-                dev_model.draw_controls(viewer_model.panel_width, viewer_model.panel_y,
+                dev_model->draw_controls(viewer_model.panel_width, viewer_model.panel_y,
                     window, error_message, device_to_remove, viewer_model, windows_width, draw_later);
             }
             if (viewer_model.ppf.is_rendering())
             {
                 if (!std::any_of(device_models->begin(), device_models->end(),
-                    [](device_model& dm)
+                    [](const std::unique_ptr<device_model>& dm)
                 {
-                    return dm.is_streaming();
+                    return dm->is_streaming();
                 }))
                 {
                     // Stopping post processing filter rendering thread
                     viewer_model.ppf.stop();
                 }
-            }
-
-            if (device_to_remove)
-            {
-                if (auto p = device_to_remove->dev.as<playback>())
-                {
-                    ctx.unload_device(p.file_name());
-                }
-                viewer_model.syncer->remove_syncer(device_to_remove->dev_syncer);
-                device_models->erase(std::find_if(begin(*device_models), end(*device_models),
-                    [&](const device_model& other) { return get_device_name(other.dev) == get_device_name(device_to_remove->dev); }));
-                device_to_remove = nullptr;
             }
 
             ImGui::SetContentRegionWidth(windows_width);
@@ -482,6 +492,26 @@ int main(int argv, const char** argc) try
                     error_message = e.what();
                 }
             }
+
+            if (device_to_remove)
+            {
+                if (auto p = device_to_remove->dev.as<playback>())
+                {
+                    ctx.unload_device(p.file_name());
+                }
+                viewer_model.syncer->remove_syncer(device_to_remove->dev_syncer);
+                auto it = std::find_if(begin(*device_models), end(*device_models),
+                    [&](const std::unique_ptr<device_model>& other)
+                { return get_device_name(other->dev) == get_device_name(device_to_remove->dev); });
+
+                if (it != device_models->end())
+                {
+                    it->reset();
+                    device_models->erase(it);
+                }
+
+                device_to_remove = nullptr;
+            }
         }
         else
         {
@@ -505,7 +535,7 @@ int main(int argv, const char** argc) try
 
     // Stop all subdevices
     for (auto&& device_model : *device_models)
-        for (auto&& sub : device_model.subdevices)
+        for (auto&& sub : device_model->subdevices)
         {
             if (sub->streaming)
                 sub->stop(viewer_model);
