+++ tools/rosbag-inspector/CMakeLists.txt
@@ -7,26 +7,12 @@ project(RealsenseToolsRosbagInspector)
 
 if(BUILD_GRAPHICAL_EXAMPLES)
 
-    if(APPLE)
-        add_definitions(-DNOC_FILE_DIALOG_OSX)
-    elseif(UNIX)
-        find_package(PkgConfig REQUIRED)
-        pkg_check_modules(GTK3 REQUIRED gtk+-3.0)
-
-        include_directories(${GTK3_INCLUDE_DIRS})
-        link_directories(${GTK3_LIBRARY_DIRS})
-        add_definitions(${GTK3_CFLAGS_OTHER})
-
-        add_definitions(-DNOC_FILE_DIALOG_GTK)
-    elseif(WIN32)
-        add_definitions(-DNOC_FILE_DIALOG_WIN32)
-    endif()
-
     set(RS_ROSBAG_INSPECTOR_CPP
             rs-rosbag-inspector.cpp
             files_container.h
             print_helpers.h
             rosbag_content.h
+            ../../common/os.cpp
             ../../third-party/imgui/imgui.cpp
             ../../third-party/imgui/imgui_draw.cpp
             ../../third-party/imgui/imgui_impl_glfw.cpp
@@ -50,7 +36,7 @@ if(BUILD_GRAPHICAL_EXAMPLES)
         ${LZ4_INCLUDE_PATH}
         )
 
-    target_link_libraries(rs-rosbag-inspector realsense-file ${DEPENDENCIES} ${GTK3_LIBRARIES})
+    target_link_libraries(rs-rosbag-inspector realsense-file ${DEPENDENCIES})
 
     set_target_properties (rs-rosbag-inspector PROPERTIES
         FOLDER Tools
