+++ tools/rosbag-inspector/rs-rosbag-inspector.cpp
@@ -24,8 +24,8 @@
 #define NOMINMAX
 #endif
 #endif
-#define NOC_FILE_DIALOG_IMPLEMENTATION
-#include <noc_file_dialog.h>
+
+#include <os.h>
 
 #include "print_helpers.h"
 #include "rosbag_content.h"
@@ -146,7 +146,7 @@ void draw_menu_bar()
         {
             if (ImGui::MenuItem("Load File..."))
             {
-                auto ret = noc_file_dialog_open(NOC_FILE_DIALOG_OPEN, "ROS-bag\0*.bag\0", NULL, NULL);
+                auto ret = file_dialog_open(rs2::file_dialog_mode::open_file, "ROS-bag\0*.bag\0", NULL, NULL);
                 if (ret)
                 {
                     files.AddFiles({ ret });
